backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
langchain-chroma
chromadb==0.5.5
# langchain-ollama
pypdf
pymupdf
python-multipart
python-dotenv
# langchain-chroma
# chromadb
groq
langchain-groq
langchain-postgres
psycopg[binary]
sentence-transformers